added integration test for routers
