+++ tests/constants.py
@@ -1,0 +1,1 @@
+TASK_URL = "/api/v1/task/"  # Базовый endpoint для задач
