+++ tests/test_api.py
@@ -1,46 +1,16 @@
-from copy import deepcopy
 from datetime import datetime, timedelta
 
 import pytest
-import pytest_asyncio
-from httpx import ASGITransport, AsyncClient
+from httpx import AsyncClient
 
 from app.core.dtos.task_dto import TaskDTO
-from app.main import app
-
-task_url = "/api/v1/task/"  # Я делаю глобальную, потому что в будущем я буду разбивать test_api на разные модули
-# Когда будет больше сущностей
-# Для каждого эндпоинта - один модуль ->
-# в каждом модуле будет глобальная переменная для url эндпоинта своя
-
-
-@pytest_asyncio.fixture(scope="function")
-async def task_for_create() -> TaskDTO:
-    return TaskDTO(
-        title="title",
-        description="description",
-        deadline=datetime.now() + timedelta(days=1),
-        is_completed=False,
-        created_at=datetime.now(),
-    )
-
-
-@pytest_asyncio.fixture(scope="function")
-async def create_task_and_get_id(
-    async_client: AsyncClient, task_for_create: TaskDTO
-) -> str:
-    """Фикстура для создания новой задачи и отдача ее id"""
-
-    response = await async_client.post(task_url, json=task_for_create.to_json())
-    assert response.status_code == 201, "Проблема с созданием задачи"
-
-    return response.json()["id"]
+from tests.constants import TASK_URL
 
 
 @pytest.mark.asyncio
 async def test_get_all_tasks(async_client: AsyncClient):
     """Это тест, который показывает, что система минимально работает"""
-    response = await async_client.get(task_url)
+    response = await async_client.get(TASK_URL)
     assert response.status_code == 200
 
 
@@ -58,7 +28,7 @@ class TestCreateTask:
 
         task_for_create.title = ""
 
-        response = await async_client.post(task_url, json=task_for_create.to_json())
+        response = await async_client.post(TASK_URL, json=task_for_create.to_json())
         assert response.status_code == 422
 
     @pytest.mark.asyncio
@@ -71,7 +41,7 @@ class TestCreateTask:
         too_long_title = "a" * (TestCreateTask.max_title_length + 1)
         task_for_create.title = too_long_title
 
-        response = await async_client.post(task_url, json=task_for_create.to_json())
+        response = await async_client.post(TASK_URL, json=task_for_create.to_json())
         assert response.status_code == 422
 
     @pytest.mark.asyncio
@@ -84,7 +54,7 @@ class TestCreateTask:
         extreme_title = "a" * TestCreateTask.max_title_length
         task_for_create.title = extreme_title
 
-        response = await async_client.post(task_url, json=task_for_create.to_json())
+        response = await async_client.post(TASK_URL, json=task_for_create.to_json())
         assert response.status_code == 201
 
     @pytest.mark.asyncio
@@ -96,7 +66,7 @@ class TestCreateTask:
 
         task_for_create.deadline = None
 
-        response = await async_client.post(task_url, json=task_for_create.to_json())
+        response = await async_client.post(TASK_URL, json=task_for_create.to_json())
         assert response.status_code == 201
 
     @pytest.mark.asyncio
@@ -107,7 +77,7 @@ class TestCreateTask:
         # Ожидаю 422
 
         task_for_create.deadline = datetime.now() - timedelta(days=1)
-        response = await async_client.post(task_url, json=task_for_create.to_json())
+        response = await async_client.post(TASK_URL, json=task_for_create.to_json())
         assert response.status_code == 422
 
 
@@ -116,7 +86,7 @@ class TestChangeDeadline:
 
     @staticmethod
     def get_url_with_id(task_id: str | int) -> str:
-        return f"{task_url}rearrange/{task_id}"
+        return f"{TASK_URL}rearrange/{task_id}"
 
     @pytest.mark.asyncio
     async def test_deadline_in_past(
@@ -152,7 +122,7 @@ class TestTaskCompleted:
 
     @staticmethod
     def get_endpoint_url(task_id: str | int) -> str:
-        return f"{task_url}completed/{task_id}"
+        return f"{TASK_URL}completed/{task_id}"
 
     @pytest.mark.asyncio
     async def test_task_not_completed_after_creation(
@@ -162,9 +132,9 @@ class TestTaskCompleted:
     ):
         """Будет проверяться, что у задачи сразу после создания не может быть таска сразу выполнена"""
 
-        current_task_url = f"{task_url}{create_task_and_get_id}"
+        current_TASK_URL = f"{TASK_URL}{create_task_and_get_id}"
 
-        task_response = await async_client.get(current_task_url)
+        task_response = await async_client.get(current_TASK_URL)
 
         assert task_response.status_code == 200
         assert not task_response.json()["is_completed"]
